Add create app function (#245)
Moved app initialisation to a create_app function
Added tests for create_app function
Removed unused lines in cloudfoundry.py
